# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Confecciones WA",
        "description": "Documentación de la API para el sistema de Confecciones WA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/colores": {
            "get": {
                "tags": [
                    "Colores"
                ],
                "summary": "Listar colores",
                "description": "Obtiene el listado paginado de colores.",
                "operationId": "2b87f99a0289749a787f97016a9dee54",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Rojo"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colores obtenidos correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Colores"
                ],
                "summary": "Registrar color",
                "description": "Crea un nuevo color.",
                "operationId": "70a120a6bbdbd300f1f60ec62591e4e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "abbreviation",
                                    "is_active"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Azul"
                                    },
                                    "abbreviation": {
                                        "type": "string",
                                        "example": "AZ"
                                    },
                                    "is_active": {
                                        "description": "true activo, false inactivo",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Color creado correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/colores/{id}": {
            "put": {
                "tags": [
                    "Colores"
                ],
                "summary": "Actualizar color",
                "description": "Actualiza la información de un color.",
                "operationId": "1ee224ea51203dc99c1b006f1a0d75bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del color",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "abbreviation",
                                    "is_active"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rojo Oscuro"
                                    },
                                    "abbreviation": {
                                        "type": "string",
                                        "example": "ROD"
                                    },
                                    "is_active": {
                                        "description": "true activo, false inactivo",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Color actualizado correctamente"
                    },
                    "404": {
                        "description": "Color no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}